cart completed

+++ js/cart.js
@@ -30,33 +30,60 @@ function open1() {
 
 function cart_items_display() {
   let data = JSON.parse(localStorage.getItem('cart-items'));
+  let empty_cart = document.getElementsByClassName('empty-cart');
+  let cart_product = document.getElementById('cart-product');
 
-  let item_card = document.getElementById('item-card');
+  if (data.length == 0) {
+    empty_cart[0].style.display = 'block';
+    cart_product.style.display = 'none';
+  }
+  else {
+    let item_card = document.getElementById('item-card');
+    item_card.innerHTML = "";
+    empty_cart[0].style.display = 'none';
+    cart_product.style.display = 'flex';
 
-  for (let i = 0; i < data.length; i++){
+    for (let i = 0; i < data.length; i++){
     
-    let div = document.createElement('div');
+      let div = document.createElement('div');
 
   
-    div.setAttribute('class', 'product-show');
+      div.setAttribute('class', 'product-show');
 
   
-    div.innerHTML = `<img src="${data[i].image1}" alt="">
-                    <a href="">${data[i].detail}</a>
-                    <div class="product-size">Size <span>OS</span> &or;</div>
-                    <div class="cart-qty">Qty <span> 1</span> &or;</div>
+      div.innerHTML = `<img src="${data[i].poster}" alt="">
+                    <a href="">${data[i].description}</a>
+                    <div class="product-size">Size <span>M</span> &or;</div>
+                    <div class="cart-qty" id="c${i}" onclick="show_update_pop(this.id)">Qty <span> ${data[i].quantity}</span> &or;</div>
                     <div class="delete" id="${i}" onclick="delete_item(this.id)">Delete</div>
                     <div class="amount">
-                        <div class="total-amount">${data[i].price}</div>
+                        <div class="total-amount">Rs. ${data[i].price}</div>
                         <div class="move">Move To Closet</div>
                     </div>`;
     
-    item_card.appendChild(div);
+      item_card.appendChild(div);
+    }
+    
   }
+
+  
+
+  item_count();
 
 }
 
 cart_items_display();
+
+//Total Item Count
+
+function item_count() {
+  let data = JSON.parse(localStorage.getItem('cart-items'));
+  let count = document.getElementById('item-count');
+
+  count.innerHTML = `(${data.length} Item(s))`;
+
+  // console.log(data.length);
+}
 
 // DELETE ITEM FROM THE CART
 
@@ -65,13 +92,18 @@ function delete_item(id) {
 
   let item = document.getElementById(id);
 
+  item.parentElement.remove();
+
   data.splice(id, 1);
 
   localStorage.setItem('cart-items',JSON.stringify(data));
 
-  item.parentElement.remove();
+  
 
-  location.reload();
+  // location.reload();
+  cart_items_display();
+  bill_show();
+  item_count();
 
 }
 
@@ -85,8 +117,10 @@ function bill_show() {
   let total = 0
 
   for (let i = 0; i < data.length; i++){
-    total += data[i].price1;
+    total += Number(data[i].price)*data[i].quantity;
   }
+
+  console.log(total);
 
   let del = "FREE";
 
@@ -104,10 +138,110 @@ function bill_show() {
   }
 
   bag_amount.innerHTML = `Rs. ${total}`;
-  bag_del_amount.innerHTML = `Rs. ${del}`;
+
+  if (del != "FREE") {
+    bag_del_amount.innerHTML = `Rs. ${del}`;
+  }
+  else {
+    bag_del_amount.innerHTML = `${del}`;
+  }
+  
   bag_total_amount.innerHTML = `Rs. ${total_amount}`;
   
 }
 
 
 bill_show();
+
+//continue shopping if cart is empty starts ----->
+
+function continue_shopping() {
+  window.open('index.html');
+}
+
+
+
+//continue shopping if cart is empty starts ----->
+
+
+//Quantity and size update starts ---->
+
+let q = 1;
+
+function show_update_pop(id) {
+  let cont = document.getElementsByClassName('quantity-container');
+
+  let cart = JSON.parse(localStorage.getItem('cart-items'));
+
+  let quant = document.getElementsByClassName('qant');
+
+  let newInput = id.split("");
+
+  let index = newInput[1];
+
+  let data = document.getElementsByClassName('update-quantity');
+
+  data[0].setAttribute('id', `${index}`);
+
+  q = cart[index].quantity;
+
+  cont[0].style.display = 'flex';
+
+  quant[0].innerHTML = `<p>${cart[index].quantity}</p>`;
+
+  show_hide_minus();
+
+}
+
+function update_and_close(id) {
+  let cont = document.getElementsByClassName('quantity-container');
+
+  let data = JSON.parse(localStorage.getItem('cart-items'));
+
+  data[id].quantity = q;
+
+  localStorage.setItem('cart-items',JSON.stringify(data));
+
+  cont[0].style.display = 'none';
+
+  // location.reload();
+  cart_items_display();
+  bill_show();
+}
+
+
+function show_hide_minus() {
+  let decrease = document.getElementById('minus');
+
+if (q == 1) {
+  decrease.style.visibility = 'hidden';
+}
+else {
+  decrease.style.visibility = 'visible';
+}
+}
+
+
+
+function increase_quantity() {
+  let quant = document.getElementsByClassName('qant');
+
+  q++;
+
+  quant[0].innerHTML = `<p>${q}</p>`;
+
+  show_hide_minus();
+}
+
+function decrease_quantity() {
+  let quant = document.getElementsByClassName('qant');
+
+  q--;
+
+  quant[0].innerHTML = `<p>${q}</p>`;
+
+  show_hide_minus();
+}
+
+//Quantity and size update ends ---->
+
